Added catch to handle common file format
Sample sheets appear to routinely include blank lines that are comprised solely of commas in between samples. To exclude these lines, I added a check to determine if the line starts with the sep character and if so, skip adding it to data_lines to prevent corruption of the dataframe in the next function.